+++ src/rsu/models.py
@@ -16,7 +16,7 @@ class RSUAward(models.Model):
     exchange = models.CharField(max_length=10, choices=EXCHANGE_CHOICES)
     symbol = models.CharField(max_length=20)
     user = models.IntegerField()
-    goal = models.DecimalField(max_digits=20, decimal_places=0, null=True, blank=True)
+    goal = models.IntegerField(null=True, blank=True)
     award_date = models.DateField(_('Award Date'), )
     award_id = models.CharField(max_length=20)
     shares_awarded = models.DecimalField(_('Awarded'), max_digits=20, decimal_places=0, null=False)
